+++ lib/govuk_publishing_components/presenters/component_wrapper_helper.rb
@@ -151,7 +151,7 @@ module GovukPublishingComponents
         return if classes.blank?
 
         class_array = classes.split(" ")
-        unless class_array.all? { |c| c.start_with?("js-", "gem-c-", "govuk-", "app-c-", "brand--", "brand__") || c == "direction-rtl" }
+        unless class_array.all? { |c| c.start_with?("js-", "gem-c-", "govuk-", "app-c-", "brand--", "brand__", "gem-print-") || c == "direction-rtl" }
           raise(ArgumentError, "Classes (#{classes}) must be prefixed with `js-`")
         end
       end